requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
asgiref==3.8.1
    # via django
certifi==2024.8.30
    # via requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via uvicorn
cryptography==43.0.3
    # via
    #   jwcrypto
    #   pyjwt
django==5.1.3
    # via
    #   -r requirements.in
    #   django-oauth-toolkit
django-oauth-toolkit==3.0.1
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
idna==3.10
    # via requests
jwcrypto==1.5.6
    # via django-oauth-toolkit
oauthlib==3.2.2
    # via django-oauth-toolkit
psycopg2==2.9.10
    # via -r requirements.in
pycparser==2.22
    # via cffi
pyjwt[crypto]==2.9.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via -r requirements.in
requests==2.32.3
    # via
    #   -r requirements.in
    #   django-oauth-toolkit
six==1.16.0
    # via python-dateutil
sqlparse==0.5.1
    # via django
typing-extensions==4.12.2
    # via jwcrypto
urllib3==2.2.3
    # via requests
uvicorn==0.32.0
    # via -r requirements.in
